+++ packages/accordion/src/components/Accordion/index.tsx
@@ -17,15 +17,22 @@ function getChildrenOf(component: React.ReactType, children: React.ReactNode) {
 
 export const Accordion = (props: AccordionProps) => {
   const context = React.useContext(AccordionGroupContext);
-  const { isOpen = false, isDisabled = false, children, ...otherAccordionProps } = props;
+  const { isOpen = false, isDisabled = false, isTransparent = false, children, ...otherAccordionProps } = props;
 
   // check isOpen on Accordion group and preference to be given to it otherwise take the indicidual accordions.
   const isOpenValue = context.isOpen ? context.isOpen : isOpen;
   const ariaId = cuid();
 
   return (
-    <AccordionContext.Provider value={{ isOpen: isOpenValue, variant: context.variant, isDisabled, ariaId }}>
-      <AccordionWrapper isSelected={isOpenValue} variant={context.variant} {...otherAccordionProps}>
+    <AccordionContext.Provider
+      value={{ isOpen: isOpenValue, variant: context.variant, isDisabled, isTransparent, ariaId }}
+    >
+      <AccordionWrapper
+        isSelected={isOpenValue}
+        variant={context.variant}
+        isTransparent={isTransparent}
+        {...otherAccordionProps}
+      >
         <>{getChildrenOf(Accordion.Header, children)}</>
         <>{getChildrenOf(Accordion.Content, children)}</>
       </AccordionWrapper>