+++ packages/accordion/src/components/AccordionContext.ts
@@ -4,6 +4,7 @@ import { AccordionGroupVariant } from '../typings';
 export interface AccordionContextState {
   isOpen: boolean;
   isDisabled: boolean;
+  isTransparent: boolean;
   ariaId: string;
   variant: AccordionGroupVariant;
 }
@@ -11,6 +12,7 @@ export interface AccordionContextState {
 export const AccordionContext = createContext<AccordionContextState>({
   isOpen: false,
   isDisabled: false,
+  isTransparent: false,
   ariaId: '',
   variant: 'primary',
 });